requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Machine Learning
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
scikit-learn==1.3.2

# Image/Video Processing
opencv-python==4.8.1.78
Pillow==10.1.0
imageio==2.31.6
imageio-ffmpeg==0.4.9

# Data Processing
numpy==1.24.4
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: GPU support (uncomment if using CUDA)
# torch-audio==2.1.1
# torchtext==0.16.1

# Additional ML libraries
transformers==4.36.0
timm==0.9.10
albumentations==1.3.1
